requirements.txt
fastapi
uvicorn
openai
pydantic
jinja2
python-multipart
google-api-python-client
python-dotenv
pytest
httpx

